refactor: fix and update the verification service

+++ identity/services/verification_token_service.py
@@ -16,7 +16,7 @@ class VerificationTokenService:
     EXPIRED_STATUS = "Token is expired"
     SUCCESS_STATUS = "success"
     NOT_FOUND_STATUS = "Invalid token"
-    NO_USER_FOUND = "No user found"
+    NO_USER_FOUND_STATUS = "No user found"
 
     def __init__(self, type: str, user: Union[None, User, AbstractUser]) -> None:
         self.type = type
@@ -28,14 +28,16 @@ class VerificationTokenService:
     def generate_email_token(self) -> Union[str, None]:
         """Generate email token"""
         if not self.user:
-            return self.NO_USER_FOUND
+            return self.NO_USER_FOUND_STATUS
 
-        unhashed_token = f"{secrets.token_urlsafe(32)}{self.user.email}"
-        hashed_token = self._hash_token(unhashed_token)
+        email_string = self._hash_token(self.user.email)
+        plain_token = f"{secrets.token_urlsafe(32)}{email_string}"
+
+        hashed_token = self._hash_token(plain_token)
         expires_at = timezone.now() + timedelta(hours=24)
 
         if self._save(hashed_token=hashed_token, expires_at=expires_at):
-            return unhashed_token
+            return plain_token
 
         return None
 
@@ -46,30 +48,39 @@ class VerificationTokenService:
         return f"{settings.BASE_URL}{relative_url}"
 
     def verify_email_token(self, token: str):
+        """Verify email token"""
         hashed_token = self._hash_token(token)
 
-        try:
-            verification_token = VerificationToken.objects.get(token=hashed_token)
+        verification_token = VerificationToken.objects.filter(
+            token=hashed_token
+        ).first()
 
-            if (
-                verification_token.expires_at
-                and timezone.now() > verification_token.expires_at
-            ):
-                return self.EXPIRED_STATUS
-
-            verification_token.user.email_verified = True
-            verification_token.user.email_verified_at = timezone.now()
-            verification_token.user.save()
-
-            verification_token.delete()
-
-        except VerificationToken.DoesNotExist:
+        if not verification_token:
             return self.NOT_FOUND_STATUS
 
+        if (
+            verification_token.expires_at
+            and timezone.now() > verification_token.expires_at
+        ):
+            return self.EXPIRED_STATUS
+
+        if not verification_token.user:
+            return self.NO_USER_FOUND_STATUS
+
+        verification_token.user.email_verified = True
+        verification_token.user.email_verified_at = timezone.now()
+        verification_token.user.save()
+
+        verification_token.delete()
+
+        return self.SUCCESS_STATUS
+
     def _save(self, hashed_token: str, expires_at: datetime) -> VerificationToken:
+        """private method to save token"""
         return VerificationToken.objects.create(
             type=self.type, user=self.user, token=hashed_token, expires_at=expires_at
         )
 
     def _hash_token(self, token: str) -> str:
+        """private method to hash token"""
         return hashlib.sha256(token.encode()).hexdigest()
